Spheron/requirements.txt
flask
google-generativeai
flask-cors
matplotlib
gunicorn
opencv-python
imgurpython
numpy
publitio
diffusers
transformers
accelerate
peft